+++ test/all.dart
@@ -1,5 +1,8 @@
 library test.rpi_gpio;
 
+import 'package:rpi_gpio/rpi_gpio.dart';
+import 'package:test/test.dart';
+
 import 'basic_test.dart' as basic;
 import 'interrupts_test.dart' as interrupts;
 import 'read_write_test.dart' as read;
@@ -14,4 +17,12 @@ runTests() {
   basic.runTests();
   read.runTests();
   interrupts.runTests();
+
+  test('pins used', () {
+    // Ensure LEDs are off
+    pin(1, output)..value = 0;
+    pin(3, output)..value = 0;
+    // Print pin state changes
+    if (recording != null) recording.printUsage();
+  });
 }